chore: only support a single sink configuration
As we were thinking through the use-case of
supporting multiple sink configurations for an
export policy it didn't seem to make sense. It was
originally added to help people transition from
one sink to another, but that could easily be done
by copying the source configuration and creating a
new exporter policy with the new sink.

This will also simplify how status information is
provided because we only need to monitor the
health of a single sink configuration.